pyproject.toml
[project]
name = "gedcom-server"
version = "0.1.0"
description = "FastMCP3 GEDCOM Genealogy Server"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "ged4py>=0.4.0",
    "rapidfuzz>=3.0.0",
    "jellyfish>=1.0.0",
    "geonamescache>=2.0.0",
    "haversine>=2.8.0",
    "requests>=2.28.0",
    "strands-agents[anthropic]>=0.1.0",
    "python-dotenv>=1.0.0",
    # Tracing / Observability
    "arize-phoenix-otel>=0.14.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    # Semantic search (optional - only loaded if SEMANTIC_SEARCH_ENABLED=true)
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
]

[project.scripts]
gedcom-server = "gedcom_server.__main__:main"

[tool.poe.tasks]
serve = "python -m gedcom_server"
test = "pytest tests/ -v"
lint = "ruff check ."
format = "ruff format ."
typecheck = "mypy gedcom_server"
deps = "deptry ."
check = ["lint", "typecheck", "test"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line length handled separately

[tool.mypy]
python_version = "3.11"
warn_return_any = false  # Many functions return parsed data
warn_unused_ignores = true
disallow_untyped_defs = false  # Start lenient, tighten later
check_untyped_defs = true
disable_error_code = ["import-untyped"]

[tool.deptry]
ignore = ["DEP002"]  # Ignore missing dev dependencies in main code

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "mypy>=1.19.1",
    "poethepoet>=0.40.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
